--- a/crates/data_preparation/src/dataloader/common/mod.rs
+++ b/crates/data_preparation/src/dataloader/common/mod.rs
@@ -1,0 +1,8 @@
+//! src/dataloader/common/mod.rs
+//!
+//! Common utilities shared across DataLoader implementations.
+//!
+//! This module currently contains only thread-local storage, but can be
+//! extended to helper functions later.
+
+pub mod thread;