--- a/crates/data_preparation/src/dataloader/iterator/mod.rs
+++ b/crates/data_preparation/src/dataloader/iterator/mod.rs
@@ -44,8 +44,8 @@ use super::workers::WorkerControl;
 use crate::dataloader::common::thread::init_worker_rng;
 
 // Declare submodules
-//pub(crate) mod inmemory;
-//pub(crate) mod iterable;
+pub(crate) mod inmemory;
+pub(crate) mod iterable;
 
 // Constants
 const WORKER_RECV_TIMEOUT_MS: u64 = 100;